pyproject.toml
[project]
name = "dir-scraper"
version = "0.1.0"
description = "Generalized directory scraper with LLM-assisted normalization and enrichment"
authors = [{ name = "Pranav Lende" }]
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "selectolax>=0.3.17",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "tenacity>=8.2.3",
    "orjson>=3.9.10",
    "structlog>=23.2.0",
    "pyyaml>=6.0.1",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "pyarrow>=14.0.1",
    "pandas>=2.1.3",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "aiosqlite>=0.19.0",
    "tldextract>=5.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-playwright>=0.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
]

[project.scripts]
dirscrape = "scraper.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

